Move handlers_*_test.dart files into handlers/*_test.dart
