+++ app/test/frontend/handlers/custom_api_test.dart
@@ -12,12 +12,12 @@ import 'package:yaml/yaml.dart';
 import 'package:pub_dartlang_org/frontend/backend.dart';
 import 'package:pub_dartlang_org/frontend/models.dart';
 
-import '../shared/handlers_test_utils.dart';
-import '../shared/utils.dart';
+import '../../shared/handlers_test_utils.dart';
+import '../../shared/utils.dart';
+import '../mocks.dart';
+import '../utils.dart';
 
-import 'handlers_test_utils.dart';
-import 'mocks.dart';
-import 'utils.dart';
+import '_utils.dart';
 
 void tScopedTest(String name, Future func()) {
   scopedTest(name, () {
