+++ app/test/frontend/handlers/redirects_test.dart
@@ -10,10 +10,10 @@ import 'package:test/test.dart';
 
 import 'package:pub_dartlang_org/frontend/handlers/redirects.dart';
 
-import '../shared/handlers_test_utils.dart';
-import '../shared/utils.dart';
+import '../../shared/handlers_test_utils.dart';
+import '../../shared/utils.dart';
 
-import 'handlers_test_utils.dart';
+import '_utils.dart';
 
 void tScopedTest(String name, Future func()) {
   scopedTest(name, () {
