+++ app/test/frontend/handlers_test.dart
@@ -20,7 +20,7 @@ import 'package:pub_dartlang_org/shared/search_service.dart';
 import '../shared/handlers_test_utils.dart';
 import '../shared/utils.dart';
 
-import 'handlers_test_utils.dart';
+import 'handlers/_utils.dart';
 import 'mocks.dart';
 import 'utils.dart';
